requirements.txt
openai==0.27.8
fastapi==0.100.0
uvicorn==0.22.0
markdown-it-py==3.0.0
pymongo==4.4.0
python-multipart==0.0.6